Move world assembly helpers to assembly package
