+++ src/main/scala/com/boombustgroup/amorfati/engine/assembly/PostMonthPopulationTransitions.scala
@@ -1,4 +1,4 @@
-package com.boombustgroup.amorfati.engine.economics
+package com.boombustgroup.amorfati.engine.assembly
 
 import com.boombustgroup.amorfati.agents.*
 import com.boombustgroup.amorfati.agents.RegionalMigration
