+++ src/main/scala/com/boombustgroup/amorfati/engine/assembly/WorldInformalEconomy.scala
@@ -1,4 +1,4 @@
-package com.boombustgroup.amorfati.engine.economics
+package com.boombustgroup.amorfati.engine.assembly
 
 import com.boombustgroup.amorfati.config.SimParams
 import com.boombustgroup.amorfati.engine.mechanisms.InformalEconomy
