+++ src/main/scala/com/boombustgroup/amorfati/engine/flows/FlowSimulation.scala
@@ -5,6 +5,7 @@ import com.boombustgroup.amorfati.agents.*
 import com.boombustgroup.amorfati.config.SimParams
 import com.boombustgroup.amorfati.engine.*
 import com.boombustgroup.amorfati.engine.SimulationMonth.{CompletedMonth, ExecutionMonth}
+import com.boombustgroup.amorfati.engine.assembly.WorldAssemblyEconomics
 import com.boombustgroup.amorfati.engine.economics.*
 import com.boombustgroup.amorfati.engine.ledger.{CorporateBondOwnership, GovernmentBondCircuit, LedgerFinancialState, RuntimeFlowProjection}
 import com.boombustgroup.amorfati.engine.markets.CorporateBondMarket
